Generating consistent hostname in Travis's MacOS
It turns out that using `hostname <X>` to set the hostname doesn't
produce a consistent state after which one can reliably get the same
hostname out of the different socket-related hostname functions. I'm not
entirely sure why this is the case, but there are people online claiming
it had to do with flushing DNS caches and similar things. In particular
there are other people experiencing this same sort of behaviour in
python@MacOS when plugging their computers to some network.

Anyway, the de-facto workaround for this situations seems unanimously to
explicitly set the HostName system configuration setting via the scutil
tool. To make sure things are working I'm not only adding this extra
call, but also printing to the screen (just for debugging) the output
from several hostname-querying python functions and tools.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>