Re-implementing email sending in NGAS
The old infrastructure was suffering from bit rotting, lacked a few
features, and was not quite correct in how messages got composed. I've
re-implemented the current set of features using the email system module
instead, which makes the code much more readable and correct.

The only main difference in terms of interface of this new
implementation is that it doesn't offer a dataInFile parameter anymore.
Instead it is simply assumed that whenever an attachmentName parameter
is given, the intention is to send it as an attachment (which happens to
be the current situation).

This reimplementation also uses the smtplib module directly (plus the
new SMTP option in the NGAS configuration), rendering the ngamsSmtpLib
module unnecessary, so I'm removing it as well. The set of functionality
offered by this old module was limited anyway (no authentication, no
secure communication, etc), which is something we can add in the future.

Other improvements found in this new version is python3-compatible code
(a dict.has_key was being used), better logging (instead of silently
ignoring email sending errors), more lean 'if' statements, using a
semaphore as a context manager and other minor things

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>