Moving port check into its own function
This has the side-effect that the socket object is immediately
destroyed, which might or might not have to do with some failures we are
suffering in MacOS where this check passes, but the NGAS server actually
fails due to the address being already in use (which is what the check
is for in the first place!).

I've also moved the check to the very beginning of the prepExtSrv method
to for quick failure.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>