Reworking is_port_available function
This was previous part of the ngamsTestSuite class but it's actually
useful outside of it as well, so I'm putting it in the utils module
instead. I'm also not throwing an exception from within that function,
but letting it return a boolean indicating whether the port is available
or not.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>