New in-memory SMTP server for tests
When created by test cases, the server automatically starts the asyncore
loop (on which it's based) on a separate thread at creation time, and
joins the thread after it closes itself, which happens automatically at
tearDown() time. A condition is used to make sure extracting emails
happens only after an email has arrived.

Particular care is taken to avoid relying on the 3.5-specific default
value of the decode_data argument (True), which in 3.6 changed to False.
This meant that using email.message_from_bytes in 3.5 fails due to a
string being given in process_message instead of a bytes object.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>