Adding support for fmt specifiers in ref files
Sometimes the reference files need a format specifier for some of their
lines. For example, lines can be composed of fixed length columns but
with %-replacements in it (which we already support).

This new code adds support for specifying an optional format string at
the beginning of each line on the reference files, enclosed by [[ and
]]. When found, the rest of the line must be a space-separated list of
strings that will be given as values to the format specifier *after*
%-replacement.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>