Adjusting and enabling mail checks in unit tests
Now that we have our mail machinery machinery for unit tests in place we
can start using it to check that mails that are intended to be sent are
actually sent, and contain the correct information. Some adjusting of
the reference files was needed in order to cope with the latest
flexibilities introduced to the unit test framework and with how emails
themselves are composed now (i.e., correctly).

In particular the test_clone module has been reworked a bit to simplify
its logic and to incorporate the testing/asserting functionality into
the main test case class.

Also simpler configuration for sending mails has been incorporated,
removing unnecessary calculations (like getting the current username,
etc), and using always 127.0.0.1 instead of localhost.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>