Bigger diffs when using assert_ref_file
zip finishes when the shortest iterable finishes, and therefore it
showed a trimmed version of the full diff between a reference file and
the data we were passing on. We are using now zip_longest which iterates
over the longest iterable, using '' to fill the gaps in the shortest.
This produces a full diff, which is what we want at the end of the day.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>