Changing the meaning of the force parameter
In the previous implementation the force parameter didn't act as an
override to the value stored in the configuration object, but actually
as an ON/OFF switch that was always required to be ON in order to send
emails *on top* of the valued stored on the configuration object, which
would override the force parameter (i.e., 'force=True' and
cfg.getNotifActive() != 1 meant that no email was sent).

The new behavior is as expected: the force parameter overrides whatever
comes in the configuration.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>