Simplifying notify() API
Firstly, the dataInFile argument was not needed anymore (it can be
currently inferred by the presence of a non-empty attachmentName).
Secondly, I've changed the name of the dataRef parameter to "msg", which
is more meaningful. I've also adjusted all calls that specified optional
arguments to specify them as keywords for easier code reading.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>