Correctly setting high disk usage mark
The code was using 100 [GB] as a way to trigger a "disk is getting full"
warning email, but this didn't actually work in some systems where disks
are big (e.g., 500 [GB], which is not unusual today). A more correct way
of doing this then is to check the amount of available space in the disk
and set a high value accordingly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>