Adding plugin module loading tests
So far we have only counted coverage of code touched by the unit tests.
Unit tests don't cover most of the plugin modules though, so we are
missing coverage in that area.

A quick compromise to increase the coverage of plugins is to simply make
sure they can be imported correctly. This will give us at least some
indication of whether the code is compatible at some level with python
2/3, and will increase our coverage a bit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>