Using explicit order to get cloning candidates
Unit tests were sometimes failing against PostgreSQL because part of the
cloning logic relied on taking the first element from an unordered
query. It was clear from the reference files that the expected behavior
was to use the lowest version of a file in these opportunities, so I've
added a new ordering option to getFileInfoFromFileId to cater for this
specific order, and started using it from the CLONE command.

After these changes all tests pass with SQLite, and therefore I'm
assuming they will happily run against the other DB engines too.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>