Mentioning docker image in install instructions
There was previously no mention of the pre-build docker images we
distribute for NGAS in the installation instructions. This commit adds
that documentation as a new top-level item within the Installation
document, while the rest of the existing content has been moved inside
an "Installing from source" section.

This is again only documentation changes, so let's [skip ci].

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>