Proper order of candidates for cloning, take 2
This new implementation orders the results the *correct* way; that is,
by descending file ID first, and by disk **label** later. The original
behavior (that used when order=1 is given) orders by disk *ID*, but main
disks cannot be expected to have lower IDs than replica disks.

Our previous attempt to fix things was obviously wrong, as the intended
behavior in the case of only a file ID being specified for cloning has
always been to clone the latest version of that file rather than the
first, just like the rest of the commands assume the same.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>