Various cleanups on CLONE command
Firstly, the _cloneExplicit method was not used, and would have never
been used due to an "if (False and ...)" statement preventing it from
executing. I've therefore removed the if statement *and* the
_cloneExplicit method. This makes sense: the whole function was a dead
path, and even if it wasn't it is still troubling that it didn't share
the same code as _cloneExec, which is the main cloning routine. If in
the future it turns out that somebody actually needs it for a faster
cloning path we can re-design this fast path in a more clever way that
avoids code repetition.

Apart from this, there were a few other minor details: an unused variable
being set, an overly complex if statement, and an unnecessary try/except
block in _clone.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>