Taking code out of the try/except block
Most of the code inside this try/except block didn't need to be there,
and the fact that it was made the block unnecessarily large and complex
to reason about. All the database selection operations are now happening
outside this block, and only the writing to the dbm file is
try/excepted.

Also, the special case of "only one file selected" was unnecessarily
complex, so I've simplified it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>