Testing bbcp in Travis
Being able to do this involves several points: firstly, we need to
retrieve, compile and install (in a visible place) bbcp. Secondly, we
need to ensure that ssh connectivity works and that bbcp works within
the confines of our testing box. bbcp internally connects to `hostname`
when target/source path doesn't include a hostname, and therefore we
also need to ensure that `hostname` resolves to a sensible IP address
like 127.0.0.1 or similar.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>