Unit test servers preparing volume directories
This is a big change: previously, the code in the NGAS server made the
distinction between a "simulation" and non-"simulation" run, and behaved
differently in how it scanned the volume directories when going ONLINE,
which was confusing to begin with. This also meant that most unit test
servers didn't actually exercise what real servers do (the exception
being those servers that "prepared their own volumes").

When in "simulation" mode, the specified online plug-in was *not*
executed, and instead a routine very similar, but slightly different, to
the one found in the generic online plug-in was executed. A similar
situation happened with the offline plug-in, which wasn't executed when
running in "simulation" mode. The generic offline plug-in is a no-op
function anyway, so no differences were expected there.

Given the scenario above, I decided to remove the checks on the
simulation flag, centralize the scanning of volume directories into the
generic online plugin, and use the generic online and offline plug-ins
during testing. Now volume directories of the unit test servers are set
up by the unit test routines every time before a server is started. The
server then uses the generic online plugin, which scans these volumes
and appends them to the list of known volumes. This works just like the
simulation-mode volume discovery logic worked, but only with one caveat:
the generic online plug-in assumes the slot ID is equivalent to the
basename of the volume within the volumes directory, while the
simulation-mode logic constructed this basename using the form
"<storageSetID>-<diskType>-<slotID>". Thus, in order to align the tests
and their expectations I had to re-define the slot IDs of the volumes
defined in the test configuration file to match the volume names that
are produced and expected by the tests.

I also took the opportunity to give a more meaningful value to the disk
type set on each test volume at preparation time, so they read "Main" or
"Rep" instead of "TEST-TYPE" like they did before.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>