Normal sorting instead of special "logical" sorting
Slot IDs are strings, and therefore this special treatment they were
given for sorting (basically accounting the case when they were simple
integers, e.g. when in "simulation" mode) was borderline useful. Sorting
them as strings is definitely the way to go; if users for some reason
wanted to have slot IDs be numbers, and have them numerically sorted,
they can prefix them with zeros.

This was also the only place where the logicalStrListSort function was
used, so let's remove it please.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>