Removing usage of ngamsCfgNoStreams.xml file
This file was basically a duplicate of ngamsCfg.xml, but without the
<Streams> element. However, every time we changed ngamsCfg.xml the
corresponding change (regardless of whether it was needed or not)
wouldn't necessarily make it into the *NoStreams.xml file. This lack of
synchronization between the files meant that until now we have
actually left a few things out from the *NoStreams version, like using
an arbitrary filesystem for testing, and other details.

This change completely removes the *NoStreams configuration file.
Instead, a configuration file is generated on the fly without a
<Streams> element *from* the ngamsCfg.xml file into the temporary
testing area. This removes the problem of getting out of sync, and takes
a step into the right direction of using a single configuration file for
testing.

The latest problem that triggered the decision of finally removing this
file was a lack of synchronization of the changes needed for
centralizing the volume directory scan during testing. Because I locally
use a Linux machine for development, and now that the simulation mode
doesn't exist anymore and the online plugins are always executed, the
tests using the NoStreams file passed fine in Linux machine, where the
modules being loaded by the linux online plugin actually exist more or
less ubiquitously. However, these tests failed in the Travis MacOS job,
where loading modules via "module load" is not even a thing. Instead of
adjusting the NoStreams configuration file to use the generic
online/offline plugins like its ngamsCfg.xml counterpart is doing now,
I decided to finally put an ends to its existence.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>