New routine to safely restart the last server
The logic within in this function was previously duplicated across a few
test cases, which one way or another tried to do exactly the same thing.
The new code caters for all these use cases, allowing users to specify
an action to be executed after the server has finished and before the
new one has started.

As part of this change I've also changed how the termExtSrv method is
signaled to keep or remove the NGAS root directory. It previously
checked at runtime the _noCleanUp flag via the getNoCleanUp() method,
which required us to set the flag to False, then call termExtSrv, then
set the flag to its previous value, all of which was a bit convoluted.
Now termExtSrv has a new keep_root argument, which defaults to the value
of the _noCleanUp flag, and which we hardcode to True when calling this
method from restart_last_server, which is simpler and reads better.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>