Don't let test sockets linger for too long
We have seen already in a few occasions that starting up a unit test
server in MacOS fails because its "address is in use". This is triggered
most easily when starting a cluster of servers, which I would guess adds
more pressure to the system. In any case, this is clearly caused by the
check we do on the port availability before starting the actual server.
Somehow the socket, which we explicitly close after bind(), is still
seen by the system. I couldn't find any documentation about this
apparent difference between MacOS and Linux, but at least it seems clear
that when enabling the SO_LINGER option a call to close() should return
immediately and should not defer the closing of the socket to a
background task on the kernel.

This solution seems to be working correctly, so we'll leave it in. This
method is currently used only at unit test time anyway, so setting this
option in this particular socket is harmless.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>