Tool refactoring: ngasArchiveClient -> fs_monitor
This is basically a complete rewrite of the ngasArchiveClient tool,
while retaining the same philosophy. Files are moved through directories
as they transition through the different steps of the archiving requests
(queue -> archiving -> archived -> removed), and when the tool stops any
pending requests are dumped into disk and picked up again when the tool
starts again. Bad and backlogged files are also moved into their
respective "end point" directories, where they are left for users to
inspect.

There was a lot of code that was unnecessary in the old version of the
tool (e.g., logfile rotation). Also, instead of using a multi-threaded
approach I went for a multi-process one to make things more scalable. A
PID lockfile has been added too, proper command-line parsing, better
usage of the python client class, and other improvements. Also, the
original code didn't have an "archiving" directory -- it instead had a
number of locks and maps to keep track of files that had been requested
for archival, which made the code a bit more difficult to understand.

Overall this new version of the tool uses less lines of code and behaves
better in all respects. The tool is also reusable outside the real of
the command-line, with the Monitor class being usable from python code
with relative ease.

The previous version of the tool lived under the ngasUtils director,
which sees little maintenance. Since this is a purely client-side tool
I moved the new version into the ngamsPClient package, where it will be
better maintained.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>