Idle suspension tests reliability reworked
These tests relied on log statements making it to the logfile, but
that's not necessarily always the case -- it depends on the logging file
handler and how often it flushes to disk, together with OS I/O caching,
etc. Therefore instead of reading the timestamp from the logfile, which
might be missing, we return the time at which we found the server went
to idling -- those should be fairly similar and serve the same purpose.
This will hopefully stop the tests from intermittently failing in some
machines.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>