Empty server list considered as non-existent
We already considered empty server lists as a valid argument in the
first check ("if servers and (host or port)"), but right after that we
required them to be explicitly not None ("if servers is not None") which
is a bit contradictory and unhelpful. Let's switch to a easier usage.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>