Unit test for the fs_monitor tool
The new Monitor object of the fs_monitor tool is flexible enough that it
allows unit testing both in- and off-process. This unit test checks that
all the basic functionality works, and covers almost 100% of the code in
the new fs_monitor module, always checking that files end up in the
correct directories after certain operations occur.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>