Unit test for register plug-ins without parameters
This effectively reproduces the error reported by Brian Cheeseman when
ran without the patch that fixes the problem.

As part of the infrastructure needed for this test I had to write a
simple "do-nothing" generic registration plug-in, something that is
currently not offered by NGAS. This might actually be useful for Brian,
who apparently was trying to use the ngamsGenDapi module for exactly
this purpose.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>