Relax constraint on netifaces
The features we need from netifaces have probably been there from the
very beginning. Our dependency on a particular version is more something
of an internal baggage: while working for a different project we
contributed a patch to netifaces to help *building* it in old Linux
boxes. This patch was included in netifaces version 10.6. Since then, we
had our other project depend on >=10.6, and when we introduced this
dependency on NGAS we blindly copied it over.

Stewart Mclay pointed this out when he started trying out NGAS 11.0.2 in
his CentOS environment, which ships a system python-netifaces package
version 10.4. Functionality-wise this is more than enough for us, and
only then it became clear that the constraint was an overshot.

If any deployment breaks due to this then what needs to be fixed is the
deployment instructions, not NGAS itself, as it's the deployment's
responsibility to make sure there is a consistent environment to deploy
NGAS into.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>