More specific error message on backlog buffering
In aaaaac14a I changed the way we were raising exceptions up through the
stack when different type of errors happened. This is in general a good
thing -- in many cases we weren't able to see the real cause of an error
because the original exception was being "shadowed" by a different one.
As a side effect of this change though, we stopped being able to clearly
identify when backlog buffering happened. This was previously signaled
by the presence of the "NGAMS_WA_BUF_DATA" string in the error message
returned to the user (not extremely robust, but it worked). Since the
original exception is now being re-raised, even after backlog buffering,
that particular string didn't show up anymore in the error messages and
we stopped being able to identify whether backlog buffering happened or
not.

This commit is combining both solutions: the original exception is still
re-raised, but the user is presented with an error messages containing
the "NGAMS_WA_BUF_DATA" string as before. This brings back a client's
ability to identify these situations (and react to them), while on the
server's log we still continue to see the original cause of the error.

In the longer term it would be even better to identify these particular
situations with HTTP status codes, such that we don't have to rely on
strings being present in the error message, and which enables an easier
check.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>