Documenting new and old tools
The new ngas-fs-monitor-client command-line tool has now gotten a
mention in the changelog, which points to a new page briefly describing
I think all command-line tools installed by the different NGAS packages.
This is a nice addition to the documentation, and will hopefully grow
into something more substantial with time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>