Preventing TypeError during date comparison
The type of lastSchedule up to this point is either NoneType or str,
but on the other hand lastDelivery is a float (the type we use
internally to hold date/times). These two values are compared for order,
and while python2 happily compares strings v/s floats, python3 refuses
to do so and bombs out with a TypeError.

Because the subscription code is so convolved, it's a bit difficult to
fix this at the very source of the problem, but ensuring the correct
data representation just before the comparison is sufficient.

This problem was obviously not caught by our unit test suite because the
conditions to trigger it were not coded into any of them. I'm thus
adding a new unit test that reproduces the original problem, and that
will ensure this doesn't happen again.

This patch solves #6.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>