Correct checksum plugin recording during backlog
The logic we were using to record the checksum plugin during backlog
buffered file archival was wrong, as it was storing the checksum plugin
used to checksum the incoming file as an integer rather than a string.
This in turn was because we were storing the checksum variant (an
integer) instead of the associated checksum name in the column, which is
defined as a varchar, and has historically recorded the 'ngamsCrc32'
value. Most database drivers happily coped with this integer value and
stored it as a string, but the sybase driver was less lenient and caught
this error.

With this change we now store the correct value in the column ('crc32'
instead of 0). Not many people run into backlogged file archival anyway,
so these integral values in the checksum column shouldn't cause havoc in
the real world. And even if they do, they can be easily changed to their
corresponding string values.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>