Correcting test expectation
This test had a wrong expectation due to the code having being copied
over from test_basic_subscription. In the latter the subscription was
created after the original file was archived, but in the former the
subscription is created before any archiving -- therefore the failure on
retrieving version 2 of the file is not always what needs to happen.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>