Calculate remote file to transfer via bbcp
This is the core change needed by #19. So far we had always specified
the source file with a simple file path, but to fetch remote files we
need to specify them in the form [user@]host:/path/to/file. The host
part is calculated with the remote IP address of the HTTP request coming
from the client. This works under the assumption a connection in the
reverse order can be established.

Regarding the last point, bbcp seems to have options to revert the
connection flow, so the source (i.e., the NGAS client) connects to the
sink (i.e., the NGAS server). This *should* work in principle, but in a
simple test using docker containers I had trouble making it work, and
since I haven't invested more time figure this out I refrained from
adding this connection flow inversion.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>