Reverting the bbcp flow
Now the bbcp source (i.e., the NGAS client) is the one connecting to the
bbcp sink (i.e., the NGAS server). Moreover, the data transfer should
happen through the same network path used by the HTTP connection, same
network interface, as we are adding the value of the "Host" HTTP header
to the bbcp target file URL, inducing the client to contact the bbcp
sink via the same network interface.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>