Return request_db keys as strings
These are written as strings, so it's weird that the expectation is that
they are returned as bytes. In terms of backwards compatibility, this
seems to be not an issue: db.keys() is called only in one place, which
iterates over the results and issues eventually a db.get(k), which we
know works with strings.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>