Revamping manual and license embedding in C apps
Until now there was a lot of code repetition on how the different manual
pages and the license text was embedded into the NGAS C lib/apps. The
way the text was dumped was inefficient, required a separate awk script,
and relied anyway in runtime string concatenation. Also, the manual
pages for some of the applications were embedded in the shared code of
the static library, making applications bigger than necessary.

With this change we are completely changing how these pieces of text end
up embedded into the code. Firstly, all *.doc files generate
corresponding *.txt.c files, all of which implement a *() function
returning the text, which is now stored statically as a single string in
each compilation unit. This file generation is done exactly in the same
way for all files, and hence it uses a shared pattern-based Makefile
rule. Secondly, the manual pages for each application are compiled only
into their respective application, and not into the static NGAS library,
which contains only the shared license text.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>