Revert the bbcp flow, part two
Most of the last changes were fine, in the sense that they made sure the
target file contained the IP address used to contact the NGAS server.
However, the reversal of the flow was not necessary, as the default
behavior of bbcp is to have the SRC contact the SINK, which is what we
want.

The addition of the -z flag was indeed mostly experimental, and it was
later revealed that it didn't make a difference to the underlying
problem (now described more thoroughly in #21).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>