Removing warnings shown by newer compilers
There was a plethora of warnings arising from the compilation of the
client library and the ngamsArchiveClient code thanks to new compilers
being more vocal about certain issues. Most of them were string format
overflows (so I changed most sprintf calls to snprintf plus a check for
the resulting number of bytes written into the string), but there were
other minor ones too: string comparison to NULL instead of strlen(s),
plus a "private" method in the ngamsCClientLib module that was fully
commented, and nobody used, and hence was removed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>