Manually start Monitor tasks
This clearly separates the creation of the Monitor from the creation of
its internal tasks, which were previously created together with the
Monitor in one go. That made it a bit difficult to discern when the
monitor should be stopped, while now it only makes sense to call
.close() on fully built Monitor objects; otherwise they can be simply
discarded.

This change, in turn, should help us to have a better signal handling in
the command line tool, which before didn't react well when signals
arrived while the Monitor object was being created.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>