React well to signals from the very start
If we are sent a signal before the background tasks of the Monitor
object are started we still want to exit gracefully. To that end we now
setup the signal handler for SIGTERM as early as possible, and also
react to KeyboardInterrupts early on by simply ignoring them.

This and the previous commit should make the command line tool, and its
test, more stable in the long run.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>