Lazily load log definitions from XML resource
These log definitions have until now always been eagerly loaded when we
import the ngamsCore module; however there are tools that don't actually
used them, in which case the eager loading is detrimental for startup
times.

We now load this lazily. There is a chance that more than one thread
tries to load the definitions at the same time, but there is no risk at
the end, since the resulting object will be the same for all purposes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>