Remove imp-based locking
In python 3.8 we are now actually getting warnings due to the imp module
getting eventually removed. We new its usage was discouraged, but now we
are really being told in the face we should stop using it.

It turns out this was actually a symptom that we were not doing things
right the begin with: importlib.import_module already checks for
sys.modules as a shortcut, and while it was possible for us to do this
check this separately in a safe manner using the imp module, it's not
possible to do safely without imp; the proper way of doing this then is
*always* calling import_module. This actually greatly simplifies the
code, and makes it stronger against multi-threading race conditions in
python 3.3+.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>