Enrich ngamsPClient's -s argument
Instead of only allowing for the XML document to be printed on stdout,
the argument now allows an extra, optional value indicating a file where
the XML status document should be written to instead of being dumped
onto stdout. Since this flag is now not only "showing" the status
document I've also changed the long flag name from "--show-status" to
simply "--status", which I think is more descriptive.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>