Simplify ngamsPClient's usage string
If we let argparse generate a usage string for us the result is simply
too long for people to make sense of it. This commit gives a simple,
predetermined usage string that users can make sense of. If they need
more details they can always use -h for a full list of options.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>