Add Travis job to check docs building
Until now we have blindly pushed documentation changes, which might or
might not break the building of the online documentation. Even worse,
sometimes the builds in RTD produce warnings, which don't necessarily
yield an automatic email warnings users that the build didn't work as
expected.

This commit adds a new job to our Travis build matrix that simply builds
the documentation. It uses the "-W --keep-going" sphinx flags so
warnings do not go unnoticed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>