Add new write_headers methods
This new method is meant to contain what's currently duplicated inside
send_data and send_file_headers, making it the one-stop shop for sending
data headers back to the user.

The response length is given special attention because we expect every
single NGAS server response to have a Content-Length header (i.e., we
don't support chunck-encoded responses), so we ensure one is given
either via the length parameter or in the given headers. Both send_data
and send_file_headers have an explicit value for the length and
mime_type parameters, so while in principle we could make them
positional parameters we make them optional to support invocations where
these values are already set in the headers (e.g., during a proxy
request).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>