Support more flexible server setups
Until now every time we start an NGAS server during our tests, its
sqlite database is stored not under the NGAS root but under a common
directory. This path is fixed, and the only flexibility that further
server instances have is to point to the same database without
re-creating it, but not to point to a different one.

This commit changes where the sqlite database is created in different
use cases. When starting single NGAS servers, each server gets its own
sqlite file created under its root directory, and when a cluster is
created (via prepCluster()) a single database file is created.

While this adds the ability to start independent servers that are not
connected through a common database, this is true only for sqlite-based
tests. Tests using other databases will still have the limitation of
being connected to the same database instance. If full independence is
needed, those tests should run only when the database backend is sqlite.